postgres: yet more migration -- dealing with dupes better

+++ src/scripts/postgresql/migrate/export_from_rethinkdb.py
@@ -1,9 +1,12 @@
 #!/usr/bin/env python
 import os, timing
 
-def process(table):
-    timing.start(table, 'export_from_rethinkdb')
+def process(table, export=True):
     out = '/migrate/data/%s'%table
+    path_to_json = out + '/smc/%s.json'%table
+    if not export:
+        return path_to_json
+    timing.start(table, 'export_from_rethinkdb')
     if os.path.exists(out):
         os.system("rm -rf %s"%out)
     s = "time rethinkdb export --password-file /migrate/secrets/rethinkdb --format json  -d %s -c db3 -e smc.%s"%(
@@ -12,7 +15,7 @@ def process(table):
     if os.system(s):
         raise RuntimeError("error exporting from rethinkdb - %s"%table)
     timing.done(table, 'export_from_rethinkdb')
-    return out + '/smc/%s.json'%table
+    return path_to_json
 
 if __name__ == "__main__":
     for file in sys.argv[1:]:
